requirements.txt
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.8
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.4.3
MarkupSafe==1.0
mccabe==0.6.1
Pillow==5.3.0
pycodestyle==2.5.0
pylint==2.4.4
requests==2.22.0
six==1.13.0
typed-ast==1.4.0
urllib3==1.25.7
webcolors==1.8.1
Werkzeug==0.14.1
wrapt==1.11.2
WTForms==2.2.1
